requirements-desktop.txt
PyQt5>=5.15.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
sentence-transformers>=3.0.0
pinecone>=3.0.0
python-dotenv>=1.0.0
packaging>=20.0
requests>=2.25.0
pyinstaller>=5.0.0